+++ packages/nakama-js/client.ts
@@ -650,8 +650,8 @@ export class Client {
   }
 
   /** A socket created with the client's configuration. */
-  createSocket(useSSL = false, verbose: boolean = false, adapter : WebSocketAdapter = new WebSocketAdapterText(), sendTimeoutSec : number): Socket {
-    return new DefaultSocket(this.host, this.port, useSSL, verbose, adapter, sendTimeoutSec);
+  createSocket(useSSL = false, verbose: boolean = false, adapter : WebSocketAdapter = new WebSocketAdapterText(), sendTimeoutMs : number = DefaultSocket.DefaultSendTimeoutMs): Socket {
+    return new DefaultSocket(this.host, this.port, useSSL, verbose, adapter, sendTimeoutMs);
   }
 
   /** Delete one or more users by ID or username. */